Split MISP module for better readability
